+++ database/migrations/2025_03_14_061954_create_professional_examinations_table.php
@@ -12,9 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('professional_examinations', function (Blueprint $table) {
-            $table->unsignedBigInteger('professional_examination_id')->autoIncrement();
+            $table->integer('professional_examination_id')->autoIncrement();
             $table->primary('professional_examination_id');
-            $table->unsignedBigInteger('graduate_id')->nullable();
+            $table->integer('graduate_id')->nullable();
             $table->foreign('graduate_id')->references('graduate_id')->on('graduates')->cascadeOnDelete();
             $table->string('name_of_examination', length: 100)->nullable()->index();
             $table->date('date_taken')->nullable();
